Mbtiles: Update references to adhere to renaming
